feat: 러닝 페이지 분할 및 진동 버그, 자유 모드 종료임에도 취소 메시지 뜨던거 해결
